FDG-10796 refactor
