+++ src/components/published-reports/download-report-table/download-button/download-button.spec.js
@@ -1,0 +1,3 @@
+describe('Report Generator Button', () => {
+  it('renders a link', () => {});
+});
